actions: Move part into zerver.forms.
